refactor: refactor project into workspace
