+++ Cargo.toml
@@ -1,33 +1,3 @@
-[package]
-name = "str0m-intro"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-name = "client"
-path = "src/bin/client/main.rs"
-
-[dependencies]
-anyhow = "1.0.95"
-once_cell = "1.20.2"
-reqwest = { version = "0.12.9", features = ["blocking", "json"] }
-rouille = { version = "3.6.2", features = ["ssl"] }
-rustls = "0.23.19"
-serde = "1.0.217"
-serde_json = "1.0.133"
-str0m = "0.6.3"
-systemstat = "0.2.3"
-tokio = { version = "1.41.1", features = ["full"] }
-tracing = "0.1.41"
-tracing-subscriber = "0.3.18"
-
-[dependencies.uuid]
-version = "1.11.0"
-features = [
-    "v4",                # Lets you generate random UUIDs
-    "fast-rng",          # Use a faster (but still sufficiently random) RNG
-    "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
-    "serde",
-]
+[workspace]
+members = ["client", "signaling"]
+resolver = "2"
