+++ client/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "client"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.95"
+reqwest = { version = "0.12.9", features = ["blocking", "json"] }
+signaling = { path = "../signaling" }
+tokio = { version = "1.41.1", features = ["full"] }
