+++ signaling/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "signaling"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+once_cell = "1.20.2"
+rouille = { version = "3.6.2", features = ["ssl"] }
+rustls = "0.23.19"
+serde = "1.0.217"
+serde_json = "1.0.133"
+str0m = "0.6.3"
+systemstat = "0.2.3"
+tracing = "0.1.41"
+tracing-subscriber = "0.3.18"
+
+[dependencies.uuid]
+version = "1.11.0"
+features = [
+    "v4",                # Lets you generate random UUIDs
+    "fast-rng",          # Use a faster (but still sufficiently random) RNG
+    "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
+    "serde",
+]
