+++ signaling/scripts/generate-certs.sh
