+++ signaling/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod client;
+pub mod message;
+pub mod util;
+
+pub enum WebRtcEvent {
+    Continue,
+    Disconnected,
+}
